Integrated lambda-buffers-common from #24, updated defToPat, reorganized modules w/ better naming practices, updated flake, updated build.nix
