+++ lambda-buffers-codegen/src/LambdaBuffers/CodeGen/Gen/Generator.hs
@@ -2,15 +2,15 @@
 {-# LANGUAGE OverloadedStrings, TypeFamilies, DataKinds, PolyKinds, GADTs, RankNTypes, MultiParamTypeClasses, FlexibleContexts, FlexibleInstances, AllowAmbiguousTypes #-}
 {-# OPTIONS_GHC -Wno-unticked-promoted-constructors #-}
 
-module LambdaBuffers.Gen.Generator where
+module LambdaBuffers.CodeGen.Gen.Generator where
 
 import Data.List ( nub, foldl' )
 import Control.Applicative (Alternative(..))
 import Data.Kind
 import Prettyprinter (Doc)
 
-import LambdaBuffers.Common.Match
-import LambdaBuffers.Common.Types
+import LambdaBuffers.CodeGen.Common.Match
+import LambdaBuffers.CodeGen.Common.Types
 
 data Lang where
   Rust :: Lang
