--- lambda-buffers-proto-compat/.envrc
@@ -1,1 +1,0 @@
-use flake ..#dev-proto-compat
