--- lambda-buffers-proto-compat/cabal.project
@@ -1,3 +1,0 @@
-packages: ./.
-
-tests: true
