--- lambda-buffers-proto-compat/hie.yaml
@@ -1,2 +1,0 @@
-cradle:
-  cabal:
