--- lambda-buffers-proto-compat/lambda-buffers-proto-compat.cabal
@@ -1,102 +1,0 @@
-cabal-version: 3.0
-name:          lambda-buffers-proto-compat
-version:       0.1.0.0
-synopsis:      Lambda Buffers Protobuf Compat. Defines a Haskell data type corresponding to a checked & validated protobuf and provides functions for translating between the proto and the Haskell type.
-
--- license:
-
-author:        MLabs LTD
-maintainer:    info@mlabs.city
-
--- A copyright notice.
--- copyright:
--- category:
-
-flag dev
-  description: Enable non-strict compilation for development
-  manual:      True
-
-common common-language
-  ghc-options:
-    -Wall -Wcompat -fprint-explicit-foralls -fprint-explicit-kinds
-    -fwarn-missing-import-lists -Weverything -Wno-unsafe
-    -Wno-missing-safe-haskell-mode -Wno-implicit-prelude
-    -Wno-missing-kind-signatures -Wno-all-missed-specializations
-
-  if !flag(dev)
-    ghc-options: -Werror
-
-  default-extensions:
-    NoStarIsType
-    BangPatterns
-    BinaryLiterals
-    ConstrainedClassMethods
-    ConstraintKinds
-    DataKinds
-    DeriveAnyClass
-    DeriveDataTypeable
-    DeriveFoldable
-    DeriveFunctor
-    DeriveGeneric
-    DeriveLift
-    DeriveTraversable
-    DerivingStrategies
-    DoAndIfThenElse
-    EmptyCase
-    EmptyDataDecls
-    EmptyDataDeriving
-    ExistentialQuantification
-    ExplicitForAll
-    ExplicitNamespaces
-    FlexibleContexts
-    FlexibleInstances
-    ForeignFunctionInterface
-    GADTSyntax
-    HexFloatLiterals
-    ImportQualifiedPost
-    InstanceSigs
-    KindSignatures
-    LambdaCase
-    MonomorphismRestriction
-    MultiParamTypeClasses
-    NamedFieldPuns
-    NamedWildCards
-    NumericUnderscores
-    OverloadedStrings
-    PartialTypeSignatures
-    PatternGuards
-    PolyKinds
-    PostfixOperators
-    RankNTypes
-    RelaxedPolyRec
-    ScopedTypeVariables
-    StandaloneDeriving
-    StandaloneKindSignatures
-    TemplateHaskell
-    TraditionalRecordSyntax
-    TupleSections
-    TypeApplications
-    TypeFamilies
-    TypeOperators
-    TypeSynonymInstances
-    ViewPatterns
-
-  default-language:   Haskell2010
-
-library
-  import:          common-language
-  build-depends:
-    , base           >=4.16
-    , text           >=1.2
-    , transformers   >=0.5
-    , lens
-    , generic-lens
-    , proto-lens
-    , template-haskell
-    , th-abstraction
-    , lambda-buffers-compiler-pb
-
-  hs-source-dirs:  src
-  exposed-modules:
-    LambdaBuffers.ProtoCompat.ProtoCompat
-    LambdaBuffers.ProtoCompat.TH
